[MODORDERS-423] - Move spring test configuration
